+++ src/main/java/chip/ChipException.java
@@ -1,3 +1,5 @@
+package chip;
+
 public class ChipException extends Exception {
     public ChipException(String message) {
         super(message);
