+++ src/main/java/chip/command/Command.java
@@ -1,3 +1,5 @@
+package chip.command;
+
 public enum Command {
     TODO, DEADLINE, EVENT, LIST, MARK, UNMARK, DELETE, BYE
 }
