+++ src/main/java/chip/command/Parser.java
@@ -1,90 +1,49 @@
-import java.util.Scanner;
-import java.util.ArrayList;
+package chip.command;
 
-public class Chip {
+import chip.task.*;
+import chip.ui.*;
+import chip.storage.*;
+import chip.ChipException;
 
-    private Storage storage;
-    private TaskList tasks;
-    private Ui ui;
+public class Parser {
 
-    public Chip(String filePath) {
-        ui = new Ui();
-        storage = new Storage(filePath);
-        try {
-            tasks = new TaskList(storage.load());
-        } catch (ChipException e) {
-            ui.showError("Data file not found. Starting with an empty task list.");
-            tasks = new TaskList();
-        }
-    }
-
-    public void run() {
-        ui.showWelcome();
-
-        while (true) {
-            try {
-                String fullCommand = ui.readCommand();
-                ui.showLine();
-
-                // Check for bye command separately since it doesn't modify tasks
-                if (fullCommand.trim().equalsIgnoreCase("bye")) {
-                    ui.showGoodbye();
-                    ui.showLine();
-                    break;
-                }
-
-                // Parse and execute command
-                executeCommand(fullCommand);
-
-            } catch (ChipException e) {
-                ui.showError(e.getMessage());
-            } catch (IllegalArgumentException e) {
-                ui.showError("I'm sorry, there is no such action.");
-            } catch (Exception e) {
-                ui.showError("An unexpected error occurred. Please check your command.");
-            } finally {
-                ui.showLine();
-            }
-        }
-    }
-
-    private void executeCommand(String fullCommand) throws ChipException {
+    public static void parse(String fullCommand, TaskList tasks, Ui ui, Storage storage) throws ChipException {
         String[] parts = fullCommand.split(" ", 2);
         Command action = Command.valueOf(parts[0].toUpperCase());
 
         switch (action) {
             case LIST:
-                showTaskList();
+                showTaskList(tasks, ui);
                 break;
             case MARK:
-                markTask(parts);
+                markTask(parts, tasks, ui, storage);
                 break;
             case UNMARK:
-                unmarkTask(parts);
+                unmarkTask(parts, tasks, ui, storage);
                 break;
             case DELETE:
-                deleteTask(parts);
+                deleteTask(parts, tasks, ui, storage);
                 break;
             case TODO:
-                addTodo(parts);
+                addTodo(parts, tasks, ui, storage);
                 break;
             case DEADLINE:
-                addDeadline(parts);
+                addDeadline(parts, tasks, ui, storage);
                 break;
             case EVENT:
-                addEvent(parts);
+                addEvent(parts, tasks, ui, storage);
                 break;
         }
     }
 
-    private void showTaskList() {
+    private static void showTaskList(TaskList tasks, Ui ui) {
         ui.showMessage("Here are the tasks in your list:");
         for (int i = 0; i < tasks.size(); i++) {
             ui.showMessage(" " + (i + 1) + "." + tasks.getTask(i));
         }
     }
 
-    private void markTask(String[] parts) throws ChipException {
+    private static void markTask(String[] parts, TaskList tasks, Ui ui, Storage storage) throws ChipException {
         if (parts.length < 2) {
             throw new ChipException("Please specify which task to mark.");
         }
@@ -93,10 +52,10 @@ public class Chip {
         task.markAsDone();
         ui.showMessage("Nice! I've marked this task as done:");
         ui.showMessage("   " + task);
-        saveTasksToFile();
+        storage.save(tasks.getTasks());
     }
 
-    private void unmarkTask(String[] parts) throws ChipException {
+    private static void unmarkTask(String[] parts, TaskList tasks, Ui ui, Storage storage) throws ChipException {
         if (parts.length < 2) {
             throw new ChipException("Please specify which task to unmark.");
         }
@@ -105,10 +64,10 @@ public class Chip {
         task.markAsNotDone();
         ui.showMessage("OK, I've marked this task as not done yet:");
         ui.showMessage("   " + task);
-        saveTasksToFile();
+        storage.save(tasks.getTasks());
     }
 
-    private void deleteTask(String[] parts) throws ChipException {
+    private static void deleteTask(String[] parts, TaskList tasks, Ui ui, Storage storage) throws ChipException {
         if (parts.length < 2) {
             throw new ChipException("Please specify which task to delete.");
         }
@@ -117,10 +76,10 @@ public class Chip {
         ui.showMessage("Noted. I've removed this task:");
         ui.showMessage("   " + removedTask);
         ui.showMessage("Now you have " + tasks.size() + " tasks in the list.");
-        saveTasksToFile();
+        storage.save(tasks.getTasks());
     }
 
-    private void addTodo(String[] parts) throws ChipException {
+    private static void addTodo(String[] parts, TaskList tasks, Ui ui, Storage storage) throws ChipException {
         if (parts.length < 2) {
             throw new ChipException("The description of a todo cannot be empty.");
         }
@@ -129,10 +88,10 @@ public class Chip {
         ui.showMessage("Got it. I've added this task:");
         ui.showMessage("   " + newTodo);
         ui.showMessage("Now you have " + tasks.size() + " tasks in the list.");
-        saveTasksToFile();
+        storage.save(tasks.getTasks());
     }
 
-    private void addDeadline(String[] parts) throws ChipException {
+    private static void addDeadline(String[] parts, TaskList tasks, Ui ui, Storage storage) throws ChipException {
         if (parts.length < 2) {
             throw new ChipException("The description of a deadline cannot be empty.");
         }
@@ -145,10 +104,10 @@ public class Chip {
         ui.showMessage("Got it. I've added this task:");
         ui.showMessage("   " + newDeadline);
         ui.showMessage("Now you have " + tasks.size() + " tasks in the list.");
-        saveTasksToFile();
+        storage.save(tasks.getTasks());
     }
 
-    private void addEvent(String[] parts) throws ChipException {
+    private static void addEvent(String[] parts, TaskList tasks, Ui ui, Storage storage) throws ChipException {
         if (parts.length < 2) {
             throw new ChipException("The description of an event cannot be empty.");
         }
@@ -165,15 +124,7 @@ public class Chip {
         ui.showMessage("Got it. I've added this task:");
         ui.showMessage("   " + newEvent);
         ui.showMessage("Now you have " + tasks.size() + " tasks in the list.");
-        saveTasksToFile();
-    }
-
-    private void saveTasksToFile() {
-        try {
-            storage.save(tasks.getTasks());
-        } catch (ChipException e) {
-            ui.showError("Error saving tasks: " + e.getMessage());
-        }
+        storage.save(tasks.getTasks());
     }
 
     public static void main(String[] args) {
