+++ src/main/java/chip/task/Event.java
@@ -1,3 +1,6 @@
+package chip.task;
+
+import chip.ChipException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
