+++ src/main/java/chip/task/Task.java
@@ -1,3 +1,7 @@
+package chip.task;
+
+import chip.ChipException;
+
 public class Task {
     private String description;
     private boolean isDone;
