+++ src/main/java/chip/task/Todo.java
@@ -1,3 +1,7 @@
+package chip.task;
+
+import chip.ChipException;
+
 public class Todo extends Task {
 
     public Todo(String description) {
