+++ src/main/java/chip/ui/Ui.java
@@ -1,3 +1,5 @@
+package chip.ui;
+
 import java.util.Scanner;
 
 public class Ui {
